Fix circular dependency, other small issues
