+++ main.py
@@ -8,7 +8,7 @@ import json
 import time
 from flask import Flask, render_template, request
 from flask_restful import Api
-from create_physprocess import get_physics_process_name
+from utils import get_physics_process_name, get_short_name
 
 app = Flask(__name__,
             static_folder='./html/static',
@@ -107,73 +107,6 @@ def split_chained_request_name(name):
     spl = name.split('-')
     return '%s-...-%s' % (spl[0], spl[-1])
 
-
-#pylint: disable=too-many-branches,too-many-statements
-# It is ok to have many ifs in this function
-def get_short_name(name):
-    """
-    Return short name of dataset name
-    """
-    spl = name.split('_')
-    short_name = spl[0]
-
-    if 'GluGluToH' in name or 'GluGluH' in name:
-        short_name = 'GluGluToH'
-    elif 'TTTo' in name:
-        short_name = 'TTbar'
-    elif 'GluGluToPseudoScalarH' in name:
-        short_name = 'GluGluToPseudoScalarH'
-    elif 'VBFHiggs' in name:
-        short_name = 'VBFHiggs'
-    elif 'ZHiggs' in name:
-        short_name = 'ZHiggs'
-    elif 'WHiggs' in name:
-        short_name = 'WHiggs'
-    elif 'GluGluToMaxmixH' in name:
-        short_name = 'GluGluToMaxmixH'
-    elif 'GluGluToContin' in name:
-        short_name = 'GluGluToContin'
-    elif 'DiPhotonJets' in name:
-        short_name = 'DiPhotonJets'
-    elif 'JJH' in name:
-        short_name = 'JJHiggs'
-    elif 'GluGluToBulkGraviton' in name:
-        short_name = 'GluGluToBulkGraviton'
-    elif 'BulkGraviton' in name:
-        short_name = 'BulkGraviton'
-    elif short_name == 'b':
-        short_name = 'bbbar4l'
-    elif short_name == 'ST':
-        short_name = 'SingleTop'
-    elif short_name == 'QCD' and 'Flat' in name and not 'herwig' in name:
-        short_name = 'Flat QCD P8'
-    elif short_name == 'QCD' and 'Flat' in name and 'herwig' in name:
-        short_name = 'Flat QCD H7'
-    elif short_name == 'QCD' and '_Pt_' in name:
-        short_name = 'QCD P8'
-
-    if 'madgraphMLM' in name:
-        short_name += ' LO MG+P8'
-    elif 'FxFx' in name or 'amcatnlo' in name:
-        short_name += ' NLO MG+P8'
-    elif 'powheg' in name and 'pythia8' in name:
-        short_name += ' NLO PH+P8'
-    elif 'sherpa' in name:
-        short_name += ' Sherpa'
-    elif 'madgraph' in name:
-        short_name += ' LO MG+P8'
-
-    if short_name.startswith('WW'):
-        short_name = short_name.replace('WW', 'VV', 1)
-    elif short_name.startswith('WZ'):
-        short_name = short_name.replace('WZ', 'VV', 1)
-    elif short_name.startswith('ZZ'):
-        short_name = short_name.replace('ZZ', 'VV', 1)
-    elif short_name.startswith('ZW'):
-        short_name = short_name.replace('ZW', 'VV', 1)
-
-    return short_name
-#pylint: enable=too-many-branches,too-many-statements
 
 def get_user_role(username, cursor):
     """
@@ -319,28 +252,25 @@ def phys_process_page(phys_process=None):
     cursor = conn.cursor()
     sql_args = [get_physics_process_name(phys_process)[0]]
 
-    sql_query = '''SELECT 1,
+    sql_query = '''SELECT shortname,
                           dataset,
-                          campaign, 
-                          shortname,
-                          interested_pwgs,
-                          output,
+                          campaign,
                           total_events,
-                          physname,
+                          output,
                           chained_request,
-                          phys_shortname
+                          interested_pwgs
                           FROM phys_process WHERE physname = ?'''
 
     rows = cursor.execute(sql_query, sql_args)
 
-    rows = [(r[3],  # 0 Short name
+    rows = [(r[0],  # 0 Short name
              r[1],  # 1 Dataset
              r[2],  # 2 Campaign
-             r[5],  # 3 MiniAOD request
-             r[6],  # 4 MiniAOD total events
-             r[4],  # 5 Interested Pwgs
-             r[7],  # 6 physname
-             split_chained_request_name(r[8]) # 7 Short chained request prepid
+             r[3],  # 3 MiniAOD total events
+             r[4],  # 4 MiniAOD output dataset
+             r[5],  # 5 Chained request prepid
+             r[6],  # 6 Interested Pwgs
+             split_chained_request_name(r[5]) # 7 Short chained request prepid
             ) for r in rows]
 
     rows = sort_rows(rows, 5)
@@ -351,7 +281,6 @@ def phys_process_page(phys_process=None):
     conn.close()
     return render_template('phys.html',
                            phys_process_long=phys_process,
-                           phys_process=get_physics_process_name(phys_process),
                            table_rows=rows,
                            user_info=user_info)
